changing the colour and size of the gitter icon for equalisation purpose
last time I couldn't find fab fa-gitter working, so used svg element for gitter and fa element for other two icons which resulted in a little difference in the colour contrast and the size of the icons. So used fab fa for all the three which I found now, didn't use svg just to shorten the code length.